ProfileInRadarView 기능 개선
- ProfileInRadarView 다른 파일로 분리
- 말풍선 프레임을 이용해 재사용 가능하도록 리팩토링
- fandomName 네이밍을 fandomImageName으로 수정
